requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-chroma==0.0.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28 